new title image
I'm away from home right now and working on this small TV that has the dreaded OVERSCAN is quite stressful. Anyway, in this commit, I added a fancy new title image for the loading screen between dimensions. Absolutely beautiful! I also changed the wording of the elytra thing in the update log, moved all of the old stuff into one folder called `old` (no that you would see that, but it's nice for me), and made the behavior pack correctly depend on the resource pack and the resource pack not depend on the behavior pack.