new credits
In this commit, I changed some wording, added some stuff to the update-log.txt file (that were already there), tweaked the shading on the square emoji in the title texture, and as the main commit message implies, added a new credits.md file. It has the painting credits but also credits for other things as well. I also removed the painting credits file, since it would be redundant now.